Allow *CryptoContext to give a hint which type of buffer is preferred
Motivation:

We should allow the CryptoContext to give a hint which type of buffer is preferred by it to minimize memory copies. For example BoringSSL based implementation prefers a direct buffer (as it need to act on the memory address) while the Bouncycastle implementation prefers a heap based buffer as it only offer methods to act on byte[] arrays.

Modifications:

- Add isDirectBufferPreferred() method to CryptoDecryptContext and CryptoEncryptContext
- Add implementation for these methods
- Use the new methods to use the correct buffer type for the AAD buffer.

Result:

Less memory copies / allocations