Correctly detect when allocation fails and return correct (#67)
value

Motivation:

The `BoringSSL` class handles the Java->Native (BoringSSL) calls. In `EVP_HPKE_CTX_new_or_throw` it checks to ensure that a new `EVP_HPKE_CTX`  is properly allocated (in the native heap) and that a pointer to it is returned. However, it checks this value against `-1` when a failure of the `EVP_HPKE_CTX_new()` native method (from BoringSSL) return a `NULL` (or `0`) upon failure, not a `-1`. This means that a failed context creation will result in a `NULL` pointer being return and later dereferenced by native code. The same problem exists for EVP_HPKE_CTX_new().

Modifications:

Detect NULL and return -1 so we correctly handle it in the java layer

Result:

No more possible dereferencing of NULL pointer