Ensure we will call read() if nothing was decoded and AUTO_READ is not used
Motivation:

We should ensure we call `ctx.read()` if the codec received a message to decode but did not produce anything and AUTO_READ is set to false. Failing to do so might result in stales as the user will not know that more reads are needed to make progress. This is inline with what ByteToMessageDecoder does under the hood

Modifications:

- Change OHttpClientCodec and OHttpServerCodec to issue reads to make progress
- Add unit test

Result:

Ensure reads are submitted when needed and so no stales happen